+++ beanBoiBackend/src/main/java/com/beanBoi/beanBoiBackend/beanBoiBackend/models/BeanPurchase.java
@@ -2,7 +2,7 @@ package com.beanBoi.beanBoiBackend.beanBoiBackend.models;
 
 public class BeanPurchase {
     private String id;
-    private long amountRemaining;
-    private long amountPurchased;
+    private float amountRemaining;
+    private float amountPurchased;
     private Bean beansPurchased;
 }
