+++ beanBoiBackend/src/main/java/com/beanBoi/beanBoiBackend/beanBoiBackend/models/Recipe.java
@@ -7,6 +7,6 @@ public class Recipe {
     private String id;
     private String name;
     private String description;
-    private long duration;
+    private float duration;
     private float ratio;
 }
